Add styling and scroll bar
Some overall structure and more specific styling changes were made to give the page a better
look. These included font changes, width adjustments of both containers, adding a grid system
to the form-container for styling purposes, and styling the button. A scroll bar was also
added to the card-holder. This will allow the user to add as many books as they want, without
having the content overflow out of the 'bookshelf'. Note that I am still working on getting
the functionality of the delete and readBook button to work